deletion of a non existent digital object will now throw 404 error
https://zimlab.uni-graz.at/gams5/production/issue-backlog/-/issues/227

+++ src/main/java/org/zim/gamsapi/DigitalObject/DigitalObjectService.java
@@ -112,6 +112,13 @@ public class DigitalObjectService implements IDigitalObjectService {
   @Override
   @Transactional
   public void delete(DigitalObject digitalObject) {
+
+    if(!digitalObjectRepository.existsById(digitalObject.getId())){
+      String msg = String.format("Failed to delete digital object with id %s. It does not exist!", digitalObject.getId());
+      log.error(msg);
+      throw new DigitalObjectNotFoundException(msg);
+    }
+
     Set<Datastream> datastreams = datastreamRepository.findAllByDigitalObject(digitalObject);
     datastreamRepository.deleteAllByDigitalObject(digitalObject);
